% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSumScores.R
\name{calculateSumScores}
\alias{calculateSumScores}
\title{Calculate Sum Scores}
\usage{
calculateSumScores(Ising_samples)
}
\arguments{
\item{Ising_samples}{a list containing dataframes with binary responses}
}
\value{
a list containing vectors of sum scores
}
\description{
Creates a list with sum scores based on the given responses.
}
\examples{
# calculateSumScores(listContainingSumScores)
}
