% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateResponses.R
\name{simulateResponses}
\alias{simulateResponses}
\title{Simulate Responses}
\usage{
simulateResponses(
  edge_weights,
  thresholds,
  perturbation_type,
  amount_of_SDs_perturbation
)
}
\arguments{
\item{edge_weights}{a matrix containing edge weights of a network structure.}

\item{thresholds}{a vector containing node thresholds of a network structure.}

\item{perturbation_type}{a string specifying a perturbation direction. Choose between "aggrevating" (+) and "alleviating" (-).}

\item{amount_of_SDs_perturbation}{an integer specifying with how many standard deviations of the threshold distribution to perturbate the threshold values.}
}
\value{
A data.frame structure containing simulated binary responses.
}
\description{
Simulates binary responses based on the input of network structure parameters.
}
\examples{
# simulateResponses(edgeWeightMatrix, thresholdVector, "aggrevating", 2)

}
