% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSumScores.R
\name{plotSumScores}
\alias{plotSumScores}
\title{Plot Sum Scores}
\usage{
plotSumScores(healthy_sum_score_long)
}
\arguments{
\item{healthy_sum_score_long}{a data.frame containing sum scores}
}
\value{
a ggline line plot
}
\description{
plots the inputted sum scores data.frames as a lineplot with conficence intervalss
}
\examples{
# plotSumScores(LongFormatSumScoreDataframe)
}
