% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDFforPlottingAndANOVA.R
\name{prepareDFforPlottingAndANOVA}
\alias{prepareDFforPlottingAndANOVA}
\title{Prepare dataframe for Plotting and Analysis of Variance}
\usage{
prepareDFforPlottingAndANOVA(sum_Ising_samples)
}
\arguments{
\item{sum_Ising_samples}{a list containing vectors of sum scores}
}
\value{
a data.frame object in long format
}
\description{
Reshape the data to long format so that it can be plotted and used in an ANOVA
}
\examples{
# prepareDFforPlottingAndANOVA(listContainingSumScoreVectors)
}
